- Updated dependencies - moved common utilities to '#shared' folder
