+++ components/XDataTable.vue
@@ -2,7 +2,7 @@
 import type Dialog from 'primevue/dialog'
 import type DataTable from 'primevue/datatable'
 import { FilterMatchMode } from 'primevue/api';
-import camelCaseToTitle from '~/utils/camelCaseToTitle.js';
+import camelCaseToTitle from '#shared/camelCaseToTitle.js';
 
 export type ViewFieldType = 'text' | 'textarea' | 'number' | 'date' | 'boolean' | 'hidden' | 'object'
