+++ domain/requirements/Solution.ts
@@ -1,6 +1,6 @@
 import { Entity, Property } from '@mikro-orm/core';
 import { type Properties } from '../types/index.js';
-import slugify from '../../utils/slugify.js';
+import slugify from '../../shared/slugify.js';
 import { Requirement } from './Requirement.js';
 import { ReqType } from './ReqType.js';
