+++ nuxt.config.ts
@@ -1,6 +1,6 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-    compatibilityDate: '2024-07-22',
+    compatibilityDate: '2024-11-05',
     devtools: {
         enabled: process.env.NODE_ENV === 'development'
     },
