+++ package.json
@@ -36,20 +36,20 @@
     },
     "dependencies": {
         "@azure/monitor-opentelemetry": "^1.8.0",
-        "@azure/openai": "^2.0.0-beta.2",
+        "@azure/openai": "^2.0.0-beta.3",
         "@mikro-orm/core": "^6.3.13",
         "@mikro-orm/postgresql": "^6.3.13",
         "@mikro-orm/reflection": "^6.3.13",
         "@mikro-orm/seeder": "^6.3.13",
         "@sidebase/nuxt-auth": "^0.9.4",
         "@slack/web-api": "^7.7.0",
-        "@vite-pwa/nuxt": "^0.10.5",
-        "mermaid": "^11.3.0",
+        "@vite-pwa/nuxt": "^0.10.6",
+        "mermaid": "^11.4.0",
         "mitt": "^3.0.1",
-        "nuxt": "^3.13.2",
+        "nuxt": "^3.14.0",
         "nuxt-primevue": "^3.0.0",
         "nuxt-security": "^2.0.0",
-        "openai": "^4.68.4",
+        "openai": "^4.71.0",
         "pg": "^8.13.1",
         "primeflex": "^3.3.1",
         "primeicons": "^7.0.0",
@@ -64,17 +64,17 @@
         "@mikro-orm/migrations": "^6.3.13",
         "@nuxt/devtools": "^1.6.0",
         "@nuxt/test-utils": "^3.14.4",
-        "@types/node": "^22.8.1",
+        "@types/node": "^22.9.0",
         "@types/pg": "^8.11.10",
         "@types/uuid": "^10.0.0",
         "@vite-pwa/assets-generator": "^0.2.6",
         "@vue/test-utils": "^2.4.6",
-        "happy-dom": "^15.7.4",
+        "happy-dom": "^15.9.0",
         "playwright-core": "^1.48.2",
         "ts-node": "^10.9.2",
         "tsx": "^4.19.2",
         "typescript": "^5.6.3",
         "vitest": "^2.1.4",
-        "vue-tsc": "^2.1.8"
+        "vue-tsc": "^2.1.10"
     }
 }
