+++ pages/o/[organization-slug]/[solution-slug]/edit-entry.client.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import slugify from '#shared/slugify.js';
+
 useHead({ title: 'Edit Solution' })
 definePageMeta({ name: 'Edit Solution' })
