+++ pages/o/[organization-slug]/[solution-slug]/goals/context-and-objective.vue
@@ -1,4 +1,6 @@
 <script lang="ts" setup>
+import debounce from '#shared/debounce';
+
 type OutcomeViewModel = {
     id: string;
     reqId: string;
