+++ server/api/parse-requirement/index.post.ts
@@ -13,6 +13,8 @@ import {
 } from "~/domain/requirements/index.js";
 import { Belongs } from "~/domain/relations/Belongs.js";
 import { Follows } from "~/domain/relations/Follows.js";
+import groupBy from "#shared/groupBy.js";
+import dedent from '#shared/dedent.js'
 
 const bodySchema = z.object({
     solutionId: z.string().uuid(),
