+++ server/data/llm-zod-schemas/UseCaseSchema.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import dedent from "#shared/dedent";
 
 export default z.object({
     type: z.literal('UseCase'),
