+++ shared/groupBy.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { groupBy } from './groupBy';
+import groupBy from '#shared/groupBy';
 
 describe('groupBy', () => {
     it('should group items by a string key', () => {
